py/requirements.txt
grpcio-tools==1.47.0
numpy==1.18.4